viz/dash/backend/app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.1.4
numpy==1.26.3
sqlalchemy==1.4.42
python-dotenv==1.0.0
scikit-learn==1.3.2
python-multipart==0.0.6
databases[sqlite]==0.8.0
aiosqlite==0.19.0